pyproject.toml
[tool.poetry]
name = "fasthtml-nb-ext"
version = "0.1.0a"
description = ""
authors = ["Philip Nuzhnyi <philip.nuzhnyy@gmail.com>"]
readme = "README.md"
packages = [{include = "fasthtml_nb_ext"}]

[tool.poetry.dependencies]
python = "^3.10"
python-fasthtml = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
